Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,17 +1,17 @@
 // Your code here...
 #include <stdio.h>
 int main() {
-    int a,b;
+    float a,b;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%f %f %c",&a,&b,&c);
     if (c=='+') {
-        printf("%d",a+b);
+        printf("%f",a+b);
     } else if (c=='-') {
-        printf("%d",a-b);
+        printf("%f",a-b);
     } else if (c=='*') {
-        printf("%d",a*b);
+        printf("%f",a*b);
     } else if (c=='/') {
-        printf("%d",a/b);
+        printf("%f",a/b);
     } else {
         printf("Error");
     }
